Finished crosshairs and worked on moving images
I have finished fixing the issue involving the png image not having a transparent background by choosing a different image all-together. Also, I added a duck and the ability of it to move from the left side to the right side by altering the x-axis.
I have to start working on the scoreboard as well as the shooting system now, but other than that, the progress is going good.